Pass DataFile id not model instance to Celery task
The mytardisbf (BioFormats) filter this app was originally based on
was passing a DataFile model instance from the Django app to the
Celery task that added metadata to the DataFile (process_meta in this
app).

The Celery docs specifically recommend passing a primary key (or
equivalent) rather than serialising a model instance.  This is
probably to protect data integrity if a database admin (or some other
code / task) removes the database data corresponding to the model
instance while it's in transit.