Disallow multi-valued metadata, delete save_*_param()
The mytardisbf (BioFormats) filter this app was originally based on
allowed a metadata ParameterName to have multiple values (passed as a
list rather than a single value).

That's not required for this app, so remove the code and simplify
save_datafile_parameters() save_dataset_parameters() by removing
save_df_param() and save_ds_param().