Modular/migration schemas; Instrument NIFCert; ...
  Split trudat.py into a more modular arrangement, simplifying
metadata code.

  Implemented the revised NIF_certified rules from TruDat 2018-02-19
meeting.  To get NIFCert metadata, a Datafile must be uploaded via an
Instrument that has metadata marking it as NIF Certifying.  This alows
both NIF Cert and other instruments to exist in the same MyTardis
instance.  Only a .PvDatasets Bruker BioSpec ZIP file can be NIF
Cert=yes, but only if it has DICOM data.  All other DataFiles are
NIF Cert=no.  To be NIF Cert=yes, a Dataset must have exactly one
DataFile with a name ending in .PvDatasets (case insensitive), which
must be NIF Cert=yes itself.

  Allow NIF_certification_enabled=yes/no metadata to be added
Instruments.  Only add NIF_certified to .PvDatasets Datafiles (and
parent Datasets) if their source Instrument has ..._enabled="yes".

  The new metadata reduces the risk of a manually or incorrectly
placed .PvDatasets file being uploaded and certifying a Dataset that
shouldn't be certified.  It also potentially reduces the number of
Datafiles that need to be batch processed (e.g. nightly Dataset
certification / recovery check).

  Replace the call_command("loaddata"...) migration with a proper
migration that uses the new modular metadata for consistency between
the code and data.  Left a version of nifcert-trudat-schemas.json with
a early version of the new Instrument metadata in the fixtures
directory in case the migration needs to be backed out.

  Made a few other unrelated minor improvements during the process.